Update the title management for : and git add --all, refacto quiz manager, remove useless librairie
